requirements.txt
requests>=2.28
numpy>=1.21
scikit-learn>=1.0
redis>=4.0
pandas>=1.3
python-dotenv>=1.0
